final commit for the page (live commits coming soon)
added and completed medium and hard gamemodes, made the logo on header interactiv and leading back to the home page just in case of difficulties with gamemodes. added win popup page after completing the level with stats and a message. removed unnecessary pages and comments/code. cleaned stuff up too.